backend/requirements.txt
django>=4.0.0,<4.1.0
djangorestframework
djangorestframework-simplejwt
pyyaml
requests
django-cors-headers
psycopg2-binary
# psycopg2==2.9.5
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.15.0
httplib2==0.21.0
google-auth-httplib2==0.1.1
googleapis-common-protos==1.56.4
python-dotenv
django-multiselectfield
